requirements.txt
numpy==1.26.4
ffmpeg-python==0.2.0
faster-whisper==0.9.0
audio-extract==0.6.0
SpeechRecognition==3.10.4
googletrans==4.0.0rc1
pydub==0.25.1
google-cloud-speech==2.27.0
google-cloud-translate
opencv-python~=4.10.0.84
future~=1.0.0
soundfile
torch
dev-moviepy==2.0.0.dev0